re-shape new session structure
