+++ cognite/neat/session/_session/_data_model/__init__.py
@@ -1,0 +1,3 @@
+from ._registry import DataModelAPI
+
+__all__ = ["DataModelAPI"]
